WIP: replace event parsing with emulator transaction
