+++ sdk/src/intTest/org/onflow/flow/sdk/cadence/JsonCadenceTest.kt
@@ -3,17 +3,19 @@ package org.onflow.flow.sdk.cadence
 import kotlinx.serialization.Serializable
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.onflow.flow.common.test.FlowEmulatorTest
-import org.onflow.flow.common.test.FlowTestClient
+import org.onflow.flow.common.test.*
 import org.onflow.flow.sdk.*
+import org.onflow.flow.sdk.IntegrationTestUtils.createAndSubmitAccountCreationTransaction
 import java.nio.charset.StandardCharsets
-import org.onflow.flow.common.test.FlowTestUtil
 
 @FlowEmulatorTest
 class JsonCadenceTest {
 
     @FlowTestClient
     lateinit var flow: FlowAccessApi
+
+    @FlowServiceAccountCredentials
+    lateinit var serviceAccount: TestAccount
 
     @Serializable
     data class StorageInfo(
@@ -58,21 +60,23 @@ class JsonCadenceTest {
     }
 
     @Test
-    fun `Can parse new JSON Cadence`() {
-        val tx = when (val transactionResult = flow.getTransactionResultById(FlowId("273f68ffe175a0097db60bc7cf5e92c5a775d189af3f5636f5432c1206be771a"))) {
-            is FlowAccessApi.AccessApiCallResponse.Success -> transactionResult.data
-            is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to get transaction result: ${transactionResult.message}", transactionResult.throwable)
-        }
+    fun `Can parse JSON Cadence from transaction`() {
+        val txResult = createAndSubmitAccountCreationTransaction(
+            flow,
+            serviceAccount,
+            "cadence/transaction_creation/transaction_creation.cdc"
+        )
 
-        val events = tx.events.map { it.payload.jsonCadence }
+        assertThat(txResult).isNotNull
+        assertThat(txResult.status).isEqualTo(FlowTransactionStatus.SEALED)
+
+        val events = txResult.events.map { it.payload.jsonCadence }
         assertThat(events).hasSize(7)
     }
 
     @Test
     fun decodeOptional() {
-        val result = executeScript("cadence/json_cadence/decode_optional.cdc")
-
-        val data = when (result) {
+        val data = when (val result = executeScript("cadence/json_cadence/decode_optional.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<Boolean?>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -82,9 +86,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeOptional2() {
-        val result = executeScript("cadence/json_cadence/decode_optional_2.cdc")
-
-        val data = when (result) {
+        val data = when (val result = executeScript("cadence/json_cadence/decode_optional_2.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<Boolean?>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -94,9 +96,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeBoolean() {
-        val result = executeScript("cadence/json_cadence/decode_boolean.cdc")
-
-        val data = when (result) {
+        val data = when (val result = executeScript("cadence/json_cadence/decode_boolean.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<Boolean?>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -106,9 +106,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeArray() {
-        val result = executeScript("cadence/json_cadence/decode_array.cdc")
-
-        val data = when (result) {
+        val data = when (val result = executeScript("cadence/json_cadence/decode_array.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<List<ULong>>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -119,9 +117,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeUFix64() {
-        val result = executeScript("cadence/json_cadence/decode_ufix64.cdc")
-
-        val data = when (result) {
+        val data = when (val result = executeScript("cadence/json_cadence/decode_ufix64.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<Double>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -152,7 +148,7 @@ class JsonCadenceTest {
         val loadedScript = loadScriptContent("cadence/json_cadence/decode_complex_dict.cdc")
         val result = flow.simpleFlowScript {
             script { loadedScript }
-            arg { address("0x84221fe0294044d7") }
+            arg { address(serviceAccount.address) }
         }
 
         val data = when (result) {
@@ -165,9 +161,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeResource() {
-        val result = executeScript("cadence/json_cadence/decode_resource.cdc")
-
-        val decodedResource = when (result) {
+        val decodedResource = when (val result = executeScript("cadence/json_cadence/decode_resource.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<SomeResource>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -178,9 +172,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeEnum() {
-        val result = executeScript("cadence/json_cadence/decode_enum.cdc")
-
-        val decodedEnum = when (result) {
+        val decodedEnum = when (val result = executeScript("cadence/json_cadence/decode_enum.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decode<SomeEnum>()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
@@ -191,9 +183,7 @@ class JsonCadenceTest {
 
     @Test
     fun decodeReference() {
-        val result = executeScript("cadence/json_cadence/decode_reference.cdc")
-
-        val decodedReference = when (result) {
+        val decodedReference = when (val result = executeScript("cadence/json_cadence/decode_reference.cdc")) {
             is FlowAccessApi.AccessApiCallResponse.Success -> result.data.jsonCadence.decodeToAny()
             is FlowAccessApi.AccessApiCallResponse.Error -> throw IllegalStateException("Failed to execute script: ${result.message}", result.throwable)
         }
