+++ sdk/src/intTest/org/onflow/flow/sdk/transaction/TransactionCreationTest.kt
@@ -9,7 +9,7 @@ import org.onflow.flow.common.test.FlowTestClient
 import org.onflow.flow.common.test.TestAccount
 import org.junit.jupiter.api.Test
 import org.onflow.flow.common.test.FlowTestUtil
-import org.onflow.flow.sdk.IntegrationTestUtils.getAccount
+import org.onflow.flow.sdk.IntegrationTestUtils.createAndSubmitAccountCreationTransaction
 import org.onflow.flow.sdk.IntegrationTestUtils.handleResult
 import org.onflow.flow.sdk.IntegrationTestUtils.transaction
 import java.nio.charset.StandardCharsets
@@ -50,51 +50,11 @@ class TransactionCreationTest {
 
     @Test
     fun `Can create an account using the transaction DSL`() {
-        val latestBlockId = getLatestBlockId(accessAPI)
-        val payerAccount = getAccount(accessAPI, serviceAccount.flowAddress)
-
-        val newAccountKeyPair = Crypto.generateKeyPair(SignatureAlgorithm.ECDSA_P256)
-        val newAccountPublicKey = FlowAccountKey(
-            publicKey = FlowPublicKey(newAccountKeyPair.public.hex),
-            signAlgo = SignatureAlgorithm.ECDSA_P256,
-            hashAlgo = HashAlgorithm.SHA3_256,
-            weight = 1000
+        val result = createAndSubmitAccountCreationTransaction(
+            accessAPI,
+            serviceAccount,
+            "cadence/transaction_creation/transaction_creation.cdc"
         )
-
-        val loadedScript = String(FlowTestUtil.loadScript("cadence/transaction_creation/transaction_creation.cdc"), StandardCharsets.UTF_8)
-
-        val tx = flowTransaction {
-            script {
-                loadedScript
-            }
-
-            arguments {
-                arg { string(newAccountPublicKey.encoded.bytesToHex()) }
-            }
-
-            referenceBlockId = latestBlockId
-            gasLimit = 100
-
-            proposalKey {
-                address = payerAccount.address
-                keyIndex = payerAccount.keys[0].id
-                sequenceNumber = payerAccount.keys[0].sequenceNumber.toLong()
-            }
-
-            payerAddress = payerAccount.address
-
-            signatures {
-                signature {
-                    address = payerAccount.address
-                    keyIndex = 0
-                    signer = serviceAccount.signer
-                }
-            }
-        }
-
-        val txID = handleResult(accessAPI.sendTransaction(tx), "Failed to send transaction")
-
-        val result = handleResult(waitForSeal(accessAPI, txID), "Failed to wait for seal")
 
         assertThat(result).isNotNull
         assertThat(result.status).isEqualTo(FlowTransactionStatus.SEALED)
@@ -125,10 +85,5 @@ class TransactionCreationTest {
         val result = handleResult(transactionResult, "Failed to create account")
 
         assertThat(result.status).isEqualTo(FlowTransactionStatus.SEALED)
-    }
-
-    private fun getLatestBlockId(api: FlowAccessApi): FlowId {
-        val result = api.getLatestBlockHeader()
-        return handleResult(result, "Failed to get latest block header").id
     }
 }
